tests/wpt/tests/html/dom/render-blocking/element-render-blocking-024.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Unknown href causes the whole document to be blocked</title>

<link rel=expect href="#unknown" blocking="render">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_true(!!document.getElementById("last")));
    t.done();
  });
}, "unknown href causes the whole document to be blocked");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="third"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="fourth"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
